Tests/LibWeb/Text/expected/HTML/load-event-waits-for-img-added-in-DOMContentLoaded.txt

img error event
window load event
